Prettify code
