+++ Model/Reporter.py
@@ -50,35 +50,24 @@ class Reporter:
         )], dtype=self.moving_cars_amount.dtype)
         self.moving_cars_amount = np.append(self.moving_cars_amount, record)
 
-    def record_all_cars_arrive(self, time: float):
-        record = np.array([(
-            time
-        )], dtype=self.all_cars_arrive_time.dtype)
-        self.all_cars_arrive_time = np.append(self.all_cars_arrive_time, record)
-
     def record_avg_wait_time(self, avg_wait_time: float):
         record = np.array([(
             avg_wait_time
         )], dtype=self.wait_times.dtype)
         self.wait_times = np.append(self.wait_times, record)
 
-    def record_generations_best_solutions(self, fitness: float, solution: np.ndarray):
+    def record_best_solutions_scores(self, fitness: float, solution: np.ndarray):
         record = np.array([(
             fitness,
             solution
         )], dtype=self.best_solutions.dtype)
         self.best_solutions = np.append(self.best_solutions, record)
 
-    def save_all_data(self, directory):
+    def save_all_data(self, directory, experiment_id):
         # Save each array to a separate file in the specified directory
-        id = 555
-        np.save(f'{directory}/wait_times_check{id}.npy', self.wait_times)
-        np.save(f'{directory}/all_cars_arrive_time{id}.npy', self.all_cars_arrive_time)
-        np.save(f'{directory}/not_reaching_cars{id}.npy', self.not_reaching_cars)
-        np.save(f'{directory}/moving_cars_amount{id}.npy', self.moving_cars_amount)
-        np.save(f'{directory}/wait_time_punishment{id}.npy', self.wait_time_punishment)
-        np.save(f'{directory}/best_solutions{id}.npy', self.best_solutions, allow_pickle=True)
-
-    def report_training_results(self, score_history, best_score):
-        pass
-
+        np.save(f'{directory}/wait_times_check_{experiment_id}.npy', self.wait_times)
+        np.save(f'{directory}/all_cars_arrive_time_{experiment_id}.npy', self.all_cars_arrive_time)
+        np.save(f'{directory}/not_reaching_cars_{experiment_id}.npy', self.not_reaching_cars)
+        np.save(f'{directory}/moving_cars_amount_{experiment_id}.npy', self.moving_cars_amount)
+        np.save(f'{directory}/wait_time_punishment_{experiment_id}.npy', self.wait_time_punishment)
+        np.save(f'{directory}/best_solutions_{experiment_id}.npy', self.best_solutions, allow_pickle=True)
