+++ PPO/ActorNetwork.py
@@ -4,8 +4,10 @@ import torch.nn as nn
 import torch.optim as optim
 from torch.distributions.categorical import Categorical
 
+
 class ActorNetwork(nn.Module):
-    def __init__(self, n_actions, input_dims, alpha, fc1_dim=256, fc2_dim=256, chkpt_dir='..\\tmp'):
+    # TODO need to run and see it saves in the new file properly
+    def __init__(self, n_actions, input_dims, alpha, fc1_dim=256, fc2_dim=256, chkpt_dir='TrainedNetworks'):
         super(ActorNetwork, self).__init__()
 
         self.checkpoint_file = os.path.join(chkpt_dir, 'actor_torch_ppo')
@@ -22,13 +24,19 @@ class ActorNetwork(nn.Module):
         self.to(self.device)
 
     def forward(self, state):
+        """
+        Perform a forward pass through the network.
+        :param state: The input state to the network.
+        :return: A categorical distribution over actions.
+        """
         dist = self.actor(state)
         dist = Categorical(dist)
         return dist
 
     def save_checkpoint(self):
+        """Save the model's parameters to a checkpoint file."""
         T.save(self.state_dict(), self.checkpoint_file)
 
     def load_checkpoint(self):
+        """Load the model's parameters from a checkpoint file."""
         self.load_state_dict(T.load(self.checkpoint_file))
-
