+++ PPO/CriticNetwork.py
@@ -5,7 +5,8 @@ import torch.optim as optim
 
 
 class CriticNetwork(nn.Module):
-    def __init__(self, input_dims, alpha, fc1_dim=256, fc2_dim=256, chkpt_dir='..\\tmp'):
+    #TODO need to run and see it saves in the new file properly
+    def __init__(self, input_dims, alpha, fc1_dim=256, fc2_dim=256, chkpt_dir='TrainedNetworks'):
         super(CriticNetwork, self).__init__()
 
         self.checkpoint_file = os.path.join(chkpt_dir, 'critic_torch_ppo')
@@ -21,11 +22,22 @@ class CriticNetwork(nn.Module):
         self.to(self.device)
 
     def forward(self, state):
+        """
+        Perform a forward pass through the network.
+        :param state: The input state to the network.
+        :return: The estimated value of the input state.
+        """
         value = self.critic(state)
         return value
 
     def save_checkpoint(self):
+        """
+        Save the model's parameters to a checkpoint file.
+        """
         T.save(self.state_dict(), self.checkpoint_file)
 
     def load_checkpoint(self):
+        """
+        Load the model's parameters from a checkpoint file.
+        """
         self.load_state_dict(T.load(self.checkpoint_file))
