+++ PPO/PPOMemory.py
@@ -1,6 +1,5 @@
-from typing import List
-
 import numpy as np
+
 
 class PPOMemory:
     def __init__(self, batch_size):
@@ -13,11 +12,15 @@ class PPOMemory:
         self.batch_size = batch_size
 
     def generate_batches(self):
+        """
+        Generate random batches from stored experiences.
+        :return: Arrays of stored states, actions, probabilities, values, rewards, dones, and the generated batches.
+        """
         n_states = len(self.states)
         batch_start = np.arange(0, n_states, self.batch_size)
         indices = np.arange(n_states, dtype=np.int64)
         np.random.shuffle(indices)
-        batches = [indices[i:i+self.batch_size] for i in batch_start]
+        batches = [indices[i:i + self.batch_size] for i in batch_start]
 
         return np.array(self.states), \
             np.array(self.actions), \
@@ -28,6 +31,15 @@ class PPOMemory:
             batches
 
     def store_memory(self, state, action, prob, val, reward, done):
+        """
+        Store an experience in memory.
+        :param state: The observed state.
+        :param action: The action taken.
+        :param prob: The probability of the taken action.
+        :param val: The value estimate of the state.
+        :param reward: The reward received.
+        :param done: Boolean indicating whether the episode is finished.
+        """
         self.states.append(state)
         self.probs.append(prob)
         self.actions.append(action)
@@ -36,6 +48,9 @@ class PPOMemory:
         self.dones.append(done)
 
     def clear_memory(self):
+        """
+        Clear all stored experiences from memory.
+        """
         self.states.clear()
         self.probs.clear()
         self.actions.clear()
