+++ Tests/PPOTest.py
@@ -2,18 +2,18 @@ from Solvers.PPOSolver import PPOSolver
 from Model.Reporter import Reporter
 
 
-n = 8  # City width
-m = 8  # City height
-t = 40  # Max time steps
-num_cars = 350  # Number of cars per city
-num_training_cities = 1000  # Number of cities to train on
+n = 8
+m = 8
+t = 40
+num_cars = 350
+num_training_cities = 1000
 
 reporter = Reporter()
 
 solver = PPOSolver(n, m, t, reporter)
 
 if __name__ == "__main__":
-    # Train the solver
     print("Training the solver...")
     solver.train(num_training_cities, num_cars)
+    reporter.save_all_data('../ReporterData/PPO', 'PPO')
     print("Training completed.")
